pyproject.toml
[tool.poetry]
name = "aigf-backend"
version = "1.0.0"
description = "AI GF backend application"
authors = ["Denis Shurygin <main@denbro.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.111.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.31"}
pydantic-settings = "^2.3.4"
loguru = "^0.7.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
alembic = "^1.13.2"
asyncpg = "^0.29.0"
rsa = "^4.9"
aiobotocore = "^2.13.1"
gunicorn = "^22.0.0"
alembic-postgresql-enum = "^1.3.0"
openai = "^1.36.0"
pandas = "^2.2.2"
websockets = "^13.1"
numpy = "1.26.4"
broadcaster = {extras = ["redis"], version = "^0.3.1"}
confluent-kafka = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
